fix(ci): extract NotifyArchived to own file; exclude HN from link checker
wiki_worker.go hit 1509 LOC after adding NotifyArchived. Move the method
to wiki_worker_archive.go (1484 LOC remaining, under the 1500 limit).

Add news.ycombinator.com to lychee exclusions — HN consistently returns
429 on CI runners due to rate limiting, causing false link-check failures.
